Added moon position functions
I found functions to position the Moon to the second, which is useful for calculating the timings of eclipses and occultations.  
I have decided to split out these calculations to another page for determining events and occurrences so to keep this page just to determine if the atmospheric conditions are apt to permit said events to be observable.  
TODO: Add the effects of a risen Moon.