Added latitude,longitude
If the observatory is not in the list of pre-defined locations, enter the latitude and longitude parameter, latlng.  System will use the Google API to lookup the timezone and the elevation for that geolocation.
TODO: the polar coordinate calculation is off by some distance at the corners of the map.  Need to improve the calculations.