Updates for Star Gazer app
After viewing the data in the Android app, Star Gazer Forecast, some updates were added.
The University of Alaska site was dropped because the hourly forecast are either out of date or non-existent.
An error in the weather forecast was fixed. There was a bug in the astronomical twilight calculation.